BJS2-52766-Добавить тест на проверку того, что Spring Context успешно поднимается
