+++ src/test/java/school/faang/user_service/controller/recommendation/RecommendationControllerTest.java
@@ -1,6 +1,5 @@
-package school.faang.user_service.service;
+package school.faang.user_service.controller.recommendation;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -10,6 +9,7 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import school.faang.user_service.controller.RecommendationController;
 import school.faang.user_service.dto.recommendation.RecommendationDto;
+import school.faang.user_service.service.RecommendationService;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
