+++ src/test/java/school/faang/user_service/validator/recommendation/RecommendationValidationExceptionTest.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service;
+package school.faang.user_service.validator.recommendation;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
